AArch64: Align return type of rejection sampling functions
This commit aligns the return type of all 3 rejection sampling functions
to uint64_t for consistency.
In practice there is no difference as values are guaranteed to be small
integers in [0,256].

Resolves https://github.com/pq-code-package/mldsa-native/issues/703

+++ dev/aarch64_clean/src/arith_native_aarch64.h
@@ -60,11 +60,11 @@ uint64_t mld_rej_uniform_asm(int32_t *r, const uint8_t *buf, unsigned buflen,
                              const uint8_t *table);
 
 #define mld_rej_uniform_eta2_asm MLD_NAMESPACE(rej_uniform_eta2_asm)
-unsigned mld_rej_uniform_eta2_asm(int32_t *r, const uint8_t *buf,
+uint64_t mld_rej_uniform_eta2_asm(int32_t *r, const uint8_t *buf,
                                   unsigned buflen, const uint8_t *table);
 
 #define mld_rej_uniform_eta4_asm MLD_NAMESPACE(rej_uniform_eta4_asm)
-unsigned mld_rej_uniform_eta4_asm(int32_t *r, const uint8_t *buf,
+uint64_t mld_rej_uniform_eta4_asm(int32_t *r, const uint8_t *buf,
                                   unsigned buflen, const uint8_t *table);
 
 #define mld_poly_decompose_32_asm MLD_NAMESPACE(poly_decompose_32_asm)
